move footprints to separate file
